ref : ServiceRequest, RecurringOffer 더미 데이터 생성 분리
